refactor(impact): extract shared metric value parser
Move the "2,000+" -> {value, prefix, suffix} parsing out of the homepage
ImpactMetrics component into a shared util so the new Impact page cards
reuse the exact same logic.
